requirements.txt
python-telegram-bot==22.0
gTTS==2.5.1 